backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.10
authlib==1.3.0
WTForms==3.1.1
email-validator==2.1.0
gunicorn==21.2.0
Werkzeug==2.3.7
